Split views/reports.py into multiple files
