+++ wagtail/admin/urls/reports.py
@@ -1,12 +1,14 @@
 from django.urls import path
 
-from wagtail.admin.views import reports
+from wagtail.admin.views.reports.audit_logging import LogEntriesView
+from wagtail.admin.views.reports.locked_pages import LockedPagesView
+from wagtail.admin.views.reports.workflows import WorkflowTasksView, WorkflowView
 
 
 app_name = 'wagtailadmin_reports'
 urlpatterns = [
-    path('locked/', reports.LockedPagesView.as_view(), name='locked_pages'),
-    path('workflow/', reports.WorkflowView.as_view(), name='workflow'),
-    path('workflow_tasks/', reports.WorkflowTasksView.as_view(), name='workflow_tasks'),
-    path('site-history/', reports.LogEntriesView.as_view(), name='site_history'),
+    path('locked/', LockedPagesView.as_view(), name='locked_pages'),
+    path('workflow/', WorkflowView.as_view(), name='workflow'),
+    path('workflow_tasks/', WorkflowTasksView.as_view(), name='workflow_tasks'),
+    path('site-history/', LogEntriesView.as_view(), name='site_history'),
 ]
